Treasure Island: Initial push
Using the arrow keys, dodge the flying dragons as you journey to
Treasure Island. This game boothed at FunFest 2018.

FunFest is a non-profit indoor carnival held annually with the goal
of providing a safe environment for families during the evening of
Halloween. For the past seven years, I have hosted a number of
games using self-made props. This year, I decided program my own
game based on this year's theme: Treasure Island.

I know that flying dragons aren't the first, or second, or third
thing that comes to mind when thinking of "Treasure Island", but
the graphics designer for the event did such a killer job that it
just works.

Having never worked with JFrames or JPanels before, the game came
to life in a span of 18 hours: 3 hours on the menu, 2 hours on the
actual game, and 15 hours sleeping + being in lecture + commuting
+ actually deciding what kind of game to make.

Oh yeah, my younger brother and his friends thought it would be
funny if he heavily increased the speed of the flying dragons
after the carnival ended.